Add handling of post request
When a post request is sent to /admin-dashboard/slots,
a new slot object is create, and its attributes are set
onto it according to the form data.  Then the redis database
is reset for the updated info.

The create button has a jquery event listener on it which
triggers the ajax post request.