Set active to false and update redis
When a slot is selected, it is broadcast on all
sockets except the sender's socket (using socket.broadcast.emit)
so that it will add the class disabled to the input field on
everyone else's browser.

Also redis is updated.  The big thing to remember here is that dataset
will lowercase any attributes, so message.scheduleId will not work you
have to use message.scheduleid.