Add branching for slot selection
If there are no taken slots, then allow the person to take it.
If there is a slot that the person already has chosen then make their
previous slot active and their new chosen slot inactive.
If anyone chooses a slot that is already occupied then nothing happens.

I had to add a studentId attribute to the slot object.