Add a cancel button
Only if the slot is inactive and the socket id
matches the studentId.  And call makeSchedulingSlots
on page load.

This is to avoid the necessity of loading the elements
from the browser side, in the view.  The reason why I
want to avoid doing this, is that from the view and index.js
you don't have access to the socket id.

So, just make the whole page render dynamically.